requirements.txt
requests==2.31.0
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3
google-generativeai==0.8.5
boto3==1.34.0
python-dotenv==1.0.0
